+++ tests/Telplin.Core.Tests/ModuleOrNamespaceTests.fs
@@ -67,3 +67,61 @@ do ()
         """
 module A
 """
+
+[<Test>]
+let ``type in a module with ModuleSuffix is not qualified with the module name, 71`` () =
+    assertSignature
+        """
+namespace Meh.Bar
+
+[<CompilationRepresentation(CompilationRepresentationFlags.ModuleSuffix)>]
+[<RequireQualifiedAccess>]
+module Telplin =
+    type V =
+        { X : int }
+        static member Zero = { X = 0}
+"""
+        """
+namespace Meh.Bar
+
+[<CompilationRepresentation(CompilationRepresentationFlags.ModuleSuffix)>]
+[<RequireQualifiedAccess>]
+module Telplin =
+    type V =
+        {
+            X: int
+        }
+
+        static member Zero: V
+"""
+
+[<Test>]
+let ``type in a module with an implicit ModuleSuffix is not qualified with the module name, 71`` () =
+    assertSignature
+        """
+namespace Meh.Bar
+
+type Telplin = | A
+
+[<RequireQualifiedAccess>]
+module Telplin =
+    type V = { X : int }
+    let zero : V = { X = 0 }
+    let mk (v: V) : Telplin = A
+    module Inner =
+        let lst : V list * V = [], { X = 0 }
+"""
+        """
+namespace Meh.Bar
+
+type Telplin = | A
+
+[<RequireQualifiedAccess>]
+module Telplin =
+    type V = { X: int }
+    val zero: V
+    val mk: v: V -> Telplin
+
+    module Inner =
+        val lst: V list * V
+"""